Minor changes
Added some comments and changed classes name accordingly to the
assignment. Minor fix on check performed on the withdrawal, since if 100
dollars were on the checking account and 100 were withdrawn, it would
apply overdraft's fee. Added fee as a local variable to the first if
case in Withdraw method for checking account and applied (both the lines
of code are commented)